requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
supabase
jinja2
python-multipart
bcrypt
passlib
email-validator
sentry-sdk
httpx
pyyaml
python-jose
libsql-client
cryptography
